pyproject.toml
[project]
name = "rapid-api"
version = "0.1.0"
description = "RapidAPI web scraper for API categorization and analysis"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "pandas>=2.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "lxml>=4.9.0",
    "fake-useragent>=1.4.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.7.2",
    "xgboost>=3.0.5",
    "umap-learn>=0.5.9.post2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
